Fixed retention days bug.

+++ rabbit_backup/test/test_rabbit_dropbox.py
@@ -18,8 +18,7 @@ class TestPyBackup(unittest.TestCase):
     def setUp(self):
         access_token_for_testing = get_access_token_for_testing()
         self.backup_job = BackupJob(access_token_for_testing, 'backup_test', 1)
-        # noinspection PyPackageRequirements
-        self.test_file_full_path = os.path.join(current_path, 'pybackup-test-file_for_uploading')
+        self.test_file_full_path = os.path.join(current_path, 'test-file_for_uploading')
         f = open(self.test_file_full_path, 'w')
         f.write('test-data')
         f.close()
@@ -28,14 +27,15 @@ class TestPyBackup(unittest.TestCase):
         self.assertEqual('test@liguoliang.com', self.backup_job.account['email'])
 
     def test_upload(self):
-        self.backup_job.backup_and_clear_history_data(self.test_file_full_path)
+        self.backup_job.backup_and_clear_history_data([self.test_file_full_path])
 
     @unittest.skip('skip the download testing')
     def test_download(self):
         self.backup_job.rabbit_dropbox.get_file("/tmp/test-file", "ibm.pdf")
 
     def tearDown(self):
-        os.remove(self.test_file_full_path)
+        if os.path.isfile(self.test_file_full_path):
+            os.remove(self.test_file_full_path)
 
 if __name__ == '__main__':
     unittest.main()
